+++ corpus/cli_filter_tests/a.clj
@@ -4,13 +4,13 @@
 
 (defdescribe ^:on-var a-1-test
   (it "works"
-    (expect (+ 1 1))))
+    (expect (+ 10 20))))
 
 (defdescribe a-2-test
   {:in-attr-map true}
   (it "works"
-    (expect (+ 1 1))))
+    (expect (+ 10 20))))
 
 (defdescribe a-3-test
   (it "works"
-    (expect (+ 1 1))))
+    (expect (+ 10 20))))