+++ corpus/find_tests/examples.clj
@@ -3,7 +3,7 @@
    [lazytest.core :refer [describe expect it]]))
 
 (defn test-fn
-  {:lazytest/test #(expect (= 0 (test-fn 1)))}
+  {:lazytest/test #(expect (zero? (test-fn 1)))}
   [a]
   (+ a a))
 
@@ -18,6 +18,6 @@
   {:lazytest/test
    (describe "top level"
      (it "test-describe example" (expect (= 1 (test-describe 1))))
-     (it "test-describe example two" (expect (= 0 (test-describe 1)))))}
+     (it "test-describe example two" (expect (zero? (test-describe 1)))))}
   [a]
   (+ a a))